+++ src/Mvc/Controller/AppModelWrapperController.php
@@ -4,10 +4,6 @@ namespace Kyte\Mvc\Controller;
 
 class AppModelWrapperController extends ModelController
 {
-    public function hook_init() {
-        $this->requireRoles = false;
-    }
-
     // public function hook_auth() {}
 
     public function hook_prequery($method, &$field, &$value, &$conditions, &$all, &$order) {